requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
python-dotenv==0.21.1
pandas==1.5.3
requests==2.28.2
google-generativeai==0.3.1
Jinja2==3.1.2
MarkupSafe==2.1.2
itsdangerous==2.1.2
numpy==1.24.2
urllib3==1.26.15